feat(template): Move to single location projects
This change removes the use of shared projects, which cause many issues insde VS, as well as VS Code (incorrect removal, incorrect addition of files, failure to properly explain how to include multiple nuget packages, complex vscode globbing).

+++ src/SolutionTemplate/UnoSolutionTemplate.Wizard.2022/UnoSolutionWizard.cs
@@ -75,50 +75,55 @@ namespace UnoSolutionTemplate.Wizard
 		{
 			ThreadHelper.ThrowIfNotOnUIThread();
 
-			if (_dte?.Solution is Solution2 solution)
+			if (_dte?.Solution is Solution2 solution && _projectName is not null)
 			{
 				var platforms = solution.Projects.OfType<Project>().FirstOrDefault(p => p.Name == "Platforms");
 				if (platforms.Object is SolutionFolder platformsFolder)
 				{
+					var sharedProject = GetAllProjects().FirstOrDefault(p => p.Name == $"{_projectName}.Shared");
+					solution.Remove(sharedProject);
+
+					Directory.Move(Path.Combine(_targetPath, $"{_projectName}.Shared"), Path.Combine(_targetPath, _projectName));
+
 					if (_useWebAssembly)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Wasm", "Wasm.winui.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Wasm", _projectName, "Wasm.winui.netcore.vstemplate");
 
 						if (!_useWebAssemblyManifestJson)
 						{
-							var webAssemblyManifestJsonPath = Path.Combine(_targetPath, $"{_projectName}.Wasm", "manifest.webmanifest");
+							var webAssemblyManifestJsonPath = Path.Combine(_targetPath, _projectName, "manifest.webmanifest");
 							File.Delete(webAssemblyManifestJsonPath);
 						}
 					}
 
 					if (_useServer)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Server", "Server.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Server", $"{_projectName}.Server", "Server.netcore.vstemplate");
 					}
 
 					if (_useiOS || _useCatalyst || _useAndroid || _useAppKit)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Mobile", "Mobile.winui.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Mobile", _projectName, "Mobile.winui.netcore.vstemplate");
 					}
 
 					if (_useGtk)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Gtk", "SkiaGtk.winui.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Gtk", _projectName, "SkiaGtk.winui.netcore.vstemplate");
 					}
 
 					if (_useFramebuffer)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Linux.FrameBuffer", "SkiaLinuxFrameBuffer.winui.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Linux.FrameBuffer", _projectName, "SkiaLinuxFrameBuffer.winui.netcore.vstemplate");
 					}
 
 					if (_useWinUI)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Windows", "WinUI.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Windows", _projectName, "WinUI.netcore.vstemplate");
 					}
 
 					if (_useWpf)
 					{
-						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Wpf", "SkiaWpf.winui.netcore.vstemplate");
+						GenerateProject(solution, platformsFolder, $"{_projectName}.Skia.Wpf", _projectName, "SkiaWpf.winui.netcore.vstemplate");
 					}
 				}
 				else
@@ -128,7 +133,7 @@ namespace UnoSolutionTemplate.Wizard
 			}
 		}
 
-		private void GenerateProject(Solution2 solution, SolutionFolder platformsFolder, string projectFullName, string templateName)
+		private void GenerateProject(Solution2 solution, SolutionFolder platformsFolder, string projectFullName, string folderName, string templateName)
 		{
 			if (_projectName != null)
 			{
@@ -137,7 +142,7 @@ namespace UnoSolutionTemplate.Wizard
 					throw new Exception("The project name should not contain spaces");
 				}
 
-				var targetPath = Path.Combine(_targetPath, projectFullName);
+				var targetPath = Path.Combine(_targetPath, folderName);
 
 				// Duplicate the template to add custom parameters
 				var workTemplateFilePath = DuplicateTemplate(solution, templateName);
@@ -219,6 +224,14 @@ namespace UnoSolutionTemplate.Wizard
 			{
 				using var reader = new StreamReader(GetType().Assembly.GetManifestResourceStream(FindManifestFileName("NuGet-netcore.config")));
 				File.WriteAllText(nugetConfigPath, reader.ReadToEnd());
+			}
+
+			var directoryBuildPropsPath = Path.Combine(_targetPath, "..\\Directory.Build.props");
+
+			if (!File.Exists(directoryBuildPropsPath))
+			{
+				using var reader = new StreamReader(GetType().Assembly.GetManifestResourceStream(FindManifestFileName("Directory.Build-netcore.props")));
+				File.WriteAllText(directoryBuildPropsPath, reader.ReadToEnd());
 			}
 
 			OpenWelcomePage();
